requirements.txt
scikit-learn
plotly
numpy
anthropic
langchainhub
llama-index-llms-openai
llama-index
llama-index-vector-stores-postgres
llama-index-vector-stores-chroma
emoji
python-dotenv==1.0.1
langchain==0.2.1
openai==1.30.5
pinecone-client==3.0.2
playwright==1.41.2
tweepy==4.14.0
lxml==5.1.0
requests==2.31.0
beautifulsoup4 ==4.12.3
langchain_openai==0.1.8
Werkzeug==2.2.2
configparser==6.0.0
psycopg2-binary==2.9.9
ragas==0.1.1
datasets==2.17.0
requests==2.31.0
pdfplumber==0.10.4
beautifulsoup4==4.12.3
pre-commit
nomic
